+++ clojure/init-cider.el
@@ -1,0 +1,35 @@
+;; Cider integrates a Clojure buffer with a REPL
+(use-package cider
+  :ensure t
+  :pin melpa-stable
+  :init
+  (setq cider-repl-pop-to-buffer-on-connect t
+        cider-show-error-buffer t
+        cider-auto-select-error-buffer t
+        cider-repl-history-file "~/.emacs.d/cider-history"
+        cider-repl-wrap-history t
+        cider-repl-history-size 100
+        cider-repl-use-clojure-font-lock t
+        cider-docview-fill-column 70
+        cider-stacktrace-fill-column 76
+        ;; Stop error buffer from popping up while working in buffers other than REPL:
+        nrepl-popup-stacktraces nil
+        nrepl-log-messages nil
+        nrepl-hide-special-buffers t
+        cider-repl-use-pretty-printing t
+        cider-repl-result-prefix ";; => ")
+
+  :config
+  (add-hook 'cider-mode-hook #'eldoc-mode)
+  (add-hook 'cider-repl-mode-hook #'paredit-mode)
+  (add-hook 'cider-repl-mode-hook #'company-mode)
+  (add-hook 'cider-mode-hook #'company-mode)
+  (add-hook 'cider-mode-hook
+            (local-set-key (kbd "<C-return>") 'cider-eval-defun-at-point)))
+
+
+;; assume paredit mode is installed
+(add-hook 'cider-repl-mode-hook #'enable-paredit-mode)
+
+
+(provide 'init-cider)